Move CLI color palette into command-agnostic root module
