+++ src/colors.rs
@@ -2,12 +2,4 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-pub mod command;
-pub mod options;
-
-pub(crate) mod analyzer;
-pub(crate) mod colors;
-pub(crate) mod dependencies;
-pub(crate) mod item;
-pub(crate) mod orphans;
-pub(crate) mod structure;
+pub(crate) mod cli;
